This project was upgraded with external_updater.
Usage: tools/external_updater/updater.sh update external/OpenCSD
For more info, check https://cs.android.com/android/platform/superproject/main/+/main:tools/external_updater/README.md

Bug: None
Test: TreeHugger

+++ decoder/source/etmv4/trc_etmv4_stack_elem.cpp
@@ -33,6 +33,7 @@
 * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 */
 
+#include <new>
 #include "opencsd/etmv4/trc_etmv4_stack_elem.h"
 
 /* implementation of P0 element stack in ETM v4 trace*/
@@ -55,7 +56,7 @@ TrcStackElemParam *EtmV4P0Stack::createParamElem(const p0_elem_t p0_type, const 
     if (pElem)
     {
         int param_idx = 0;
-        int params_to_fill = params.size();
+        int params_to_fill = (int)params.size();
         while ((param_idx < 4) && params_to_fill)
         {
             pElem->setParam(params[param_idx], param_idx);
